requirements.txt
streamlit
folium_static
